refactor: moved detailed packages
