+++ src/main/java/kitchenpos/dao/OrderLineItemDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.OrderLineItem;
+import kitchenpos.domain.order.OrderLineItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
