+++ src/main/java/kitchenpos/dao/ProductDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
